+++ inject.js
@@ -306,9 +306,9 @@ var snuslashcommands = {
         "hint": "UnManadtory; Set all mandatory fields to false (Admin only)"
     },
     "up": {
-        "url": "sys_ui_policy_list.do?sysparm_query=nameLIKE$0^ORDERBYDESCsys_updated_on",
+        "url": "sys_ui_policy_list.do?sysparm_query=short_descriptionLIKE$0^ORDERBYDESCsys_updated_on",
         "hint": "UI Policies <name>",
-        "fields": "name"
+        "fields": "short_description,sys_updated_on"
     },
     "va": {
         "url": "/$conversation-builder.do",
@@ -349,7 +349,7 @@ var snuslashcommands = {
 var snuslashswitches = {
     "t": { "description": "View Table Structure ➚", "value": "sys_db_object.do?sys_id=$0&sysparm_refkey=name", "type": "link" },
     "n": { "description": "New Record ➚", "value": "$0.do", "type": "link" },
-    "r": { "description": "Open Random Record ➚", "value": "$random.$0", "type": "link" },
+    "r": { "description": "Open Random Record ➚", "value": "&snurandom=true", "type": "querypart" },
     "ra": { "description": "REST API Explorer ➚", "value": "$restapi.do?tableName=$0", "type": "link" },
     "c": { "description": "Table Config ➚", "value": "personalize_all.do?sysparm_rules_table=$0&sysparm_rules_label=$0", "type": "link" },
     "erd": { "description": "View Schema Map ➚", "value": "generic_hierarchy_erd.do?sysparm_attributes=table_history=,table=$0,show_internal=true,show_referenced=true,show_referenced_by=true,show_extended=true,show_extended_by=true,table_expansion=,spacing_x=60,spacing_y=90,nocontext", "type": "link" },
@@ -469,7 +469,7 @@ function snuGetDirectLinks(targeturl, shortcut) {
     var fields = "";
     var overwriteurl = "";
     try {
-        fields = (snuslashcommands[shortcut].hasOwnProperty("fields")) ? snuslashcommands[shortcut].fields || "" : "";
+        fields = (snuslashcommands[shortcut].hasOwnProperty("fields")) ? snuslashcommands[shortcut].fields || "sys_updated_on,sys_updated_by" : "sys_updated_on,sys_updated_by";
     } catch (e) { }
 
     try {
@@ -488,7 +488,7 @@ function snuGetDirectLinks(targeturl, shortcut) {
         } catch (ex) {
             return false;
         }
-        snuFetchData(g_ck, url, null, function (jsn) {
+        snuFetchData(g_ck, url, null, jsn => {
             var directlinks = '';
             if (jsn.hasOwnProperty('result')) {
                 var results = jsn.result;
@@ -525,6 +525,9 @@ function snuGetDirectLinks(targeturl, shortcut) {
                     }
                     directlinks += dispIdx + ' <a ' + idattr + '" target="' + target + '" href="' + link + '">' + txt + '</a><br />';
                 });
+                if (directlinks.length > 50) 
+                    directlinks += `<span style="opacity:0.4; font:smaller">Results: ${jsn.resultcount}</br><br />`;
+                
             }
             else {
                 directlinks = `No access to data`;
@@ -754,14 +757,7 @@ function snuSlashCommandAddListener() {
         query = query.trim();
 
 
-        targeturl = targeturl.replace(/\$0/g, query + (e.key.length == 1 ? e.key : ""));
-        if (query.split(" ").length > 0) {  //replace $1,$2 for Xth word in string
-            var queryArr = query.split(" ");
-            for (var i = 0; i <= queryArr.length; i++) {
-                var re = new RegExp("\\$" + (i + 1), "g");
-                targeturl = targeturl.replace(re, queryArr[i] || "");
-            }
-        }
+        targeturl = snuResolve$(targeturl, query, e);
         if (e.key == 'ArrowRight' || (e.key == 'Enter' && inlineOnly && !(e.ctrlKey || e.metaKey))) {
             snuSlashLog(true);
             snuGetDirectLinks(targeturl, shortcut);
@@ -993,7 +989,7 @@ function snuSlashCommandAddListener() {
                     if (typeof qry != 'undefined') {
                         if (targeturl.includes("{}") || linkSwitch) {
                             targeturl = targeturl.replace('{}', qry.getTableName());
-                            if (!targeturl.startsWith("$random")) window.open(targeturl, '_blank');
+                            if (!targeturl.includes("snurandom=true")) window.open(targeturl, '_blank');
                         }
                         else if (targeturl.startsWith("&")) {
                             var myurl = doc.location.href
@@ -1022,10 +1018,10 @@ function snuSlashCommandAddListener() {
                     else if (!linkSwitch){
                         targeturl = "/" + g_form.getTableName() + "_list.do?sysparm_query=" + targeturl;
                     }
-                    if (!targeturl.startsWith("$random"))
+                    if (!targeturl.includes("snurandom=true"))
                         window.open(targeturl, '_blank');
                 }
-                if (!targeturl.startsWith("$random")) {
+                if (!targeturl.includes("snurandom=true")) {
                     snuSlashCommandHide();
                     return;
                 }
@@ -1122,10 +1118,13 @@ function snuSlashCommandAddListener() {
 
             var inIFrame = !targeturl.startsWith("http") && !targeturl.startsWith("/") && sameWindow;
             if (e.target.className == "snutils") inIFrame = false;
-            if (targeturl.startsWith("$random")) {
+            if (targeturl.includes("snurandom=true")) {
+                let searchParams = new URLSearchParams(targeturl.split("?")[1]);
+                let q = searchParams.get("sysparm_query") || "";
+                let tableName = targeturl.split("_list.do")[0] || "notable";
                 targeturl = targeturl.substring(8)
-                snuGetRandomRecord(targeturl, "", false, res => {
-                    targeturl = targeturl + ".do?sys_id=" + res;
+                snuGetRandomRecord(tableName, q, false, res => {
+                    targeturl = tableName + ".do?sys_id=" + res;
                     if (inIFrame){
                         var nxtHdr = window.querySelectorShadowDom?.querySelectorDeep("sn-polaris-header");
                         if (nxtHdr){ //next experience
@@ -1396,6 +1395,18 @@ function snuDiffXml(shortcut, instance = '') {
     return;
 }
 
+
+function snuResolve$(targeturl, query, e) {
+    targeturl = targeturl.replace(/\$0/g, query + (e.key.length == 1 ? e.key : ""));
+    if (query.split(" ").length > 0) {  //replace $1,$2 for Xth word in string
+        var queryArr = query.split(" ");
+        for (var i = 0; i <= queryArr.length; i++) {
+            var re = new RegExp("\\$" + (i + 1), "g");
+            targeturl = targeturl.replace(re, queryArr[i] || "");
+        }
+    }
+    return targeturl;
+}
 
 function snuResolveVariables(variableString){
 
@@ -2513,7 +2524,7 @@ function snuAddTechnicalNames() {
         try {
             jQuery('h1.navbar-title div.pointerhand').css("float", "left");
             jQuery("h1.navbar-title:not(:contains('|'))").first().append('<span class="snuwrap">&nbsp;| <span style="font-family:monospace; font-size:small;">' + g_form.getTableName() +
-                ' <a onclick="snuShowScratchpad()">[scratchpad]</a><a title="Show the originating table innfo of fields" onclick="snuExtendedFieldInfo()">[table fields]</a>'+
+                ' <a onclick="snuShowScratchpad()">[scratchpad]</a><a title="Show the originating table info of fields" onclick="snuExtendedFieldInfo()">[table fields]</a>'+
                 '<a title="For easier copying of field names" onclick="snuToggleLabel()">[toggle label]</a> </span></span>');
 
             jQuery(".label-text:not(:contains('|'))").each(function (index, elem) {
@@ -3412,7 +3423,8 @@ async function snuFetchData(token, url, post, callback) {
           headers,
           body: post ? JSON.stringify(post) : null
         });
-        const data = response.ok ? await response.json() : response;
+        let data = response.ok ? await response.json() : response;
+        data.resultcount = response.headers.get("X-Total-Count");
         if (callback) callback(data);
         resolve(data);
       } catch (error) {